PySparkler: Fix logic to write files only when not being run in dry run mode.

+++ pysparkler/pysparkler/api.py
@@ -23,9 +23,12 @@ from pysparkler.pyspark_24_to_30 import visit_pyspark_24_to_30
 class PySparkler:
     """Main class for PySparkler"""
 
-    def __init__(self, from_pyspark: str = "2.4", to_pyspark: str = "3.0"):
+    def __init__(
+        self, from_pyspark: str = "2.4", to_pyspark: str = "3.0", dry_run: bool = False
+    ):
         self.from_pyspark = from_pyspark
         self.to_pyspark = to_pyspark
+        self.dry_run = dry_run
 
     def upgrade(self, input_file: str, output_file: str | None = None) -> str:
         """Upgrade the PySparkler file to the latest version and provides comments as hints for manual changes"""
@@ -37,10 +40,15 @@ class PySparkler:
         # Apply the re-writer to the AST
         modified_tree = visit_pyspark_24_to_30(original_tree)
 
-        if output_file:
-            # Write the modified AST to a file
-            with open(output_file, "w", encoding="utf-8") as f:
-                f.write(modified_tree.code)
+        if not self.dry_run:
+            if output_file:
+                # Write the modified AST to the output file location
+                with open(output_file, "w", encoding="utf-8") as f:
+                    f.write(modified_tree.code)
+            else:
+                # Re-write the modified AST back to the input file
+                with open(input_file, "w", encoding="utf-8") as f:
+                    f.write(modified_tree.code)
 
         # Verify the AST was modified
         return modified_tree.code
